% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box.Ljung.Test.R
\name{Box.Ljung.Test}
\alias{Box.Ljung.Test}
\title{Ljung-Box Test Plot}
\usage{
Box.Ljung.Test(z, lag = NULL, main = NULL)
}
\arguments{
\item{y}{(type: numeric): data vector}

\item{x}{(type: numeric): optional vector, if \code{x = NULL} then the function uses
\eqn{(1,\ldots,n)} where \code{n} is the length of \code{y}. More details in \code{y} argument
from \code{\link[graphics]{persp}} function.}

\item{n}{(type: numeric): value corresponding to the length of the window to compute periodogram.
If \code{N=NULL} then the function will use \eqn{N = \textmd{trunc}(n^{0.8})}, see
\insertCite{dahlhaus1998optimal;textual}{lsts} where \eqn{n} is the length of the \code{y} vector.}

\item{s}{(type: numeric): value corresponding to the lag with which will be taking the blocks or
windows to calculate the periodogram.}

\item{p}{(type: numeric): value used if it is desired that \code{S} is proportional to \code{N}.
By default \code{p=0.25}, if \code{S} and \code{N} are not entered.}

\item{spar.freq}{(type: numeric) smoothing parameter, typically (but not necessarily) in \eqn{(0,1]}.}

\item{spar.time}{(type: numeric) smoothing parameter, typically (but not necessarily) in \eqn{(0,1]}.}

\item{theta}{(type: numeric) angle defining the viewing direction, ives the azimuthal direction.}

\item{phi}{(type: numeric) angle defining the viewing direction, gives the colatitude.}

\item{xlim}{(type: numeric) x-axis limits, \code{NULL} by default and optional parameter.}

\item{ylim}{(type: numeric) y-axis limits, \code{NULL} by default and optional parameter.}

\item{zlim}{(type: numeric) z-axis limits, \code{NULL} by default and optional parameter.}

\item{ylab}{(type: character) title for the y-axis. By default is \code{ylab="Time"} and is an
optional parameter.}

\item{palette.col}{(type: character) colors palette.}
}
\value{
** COMPLETE **
}
\description{
Plots the p-values Ljung-Box test.
}
\details{
The Ljung-Box test is used to check if exists autocorrelation in a time series. The statistic is
\deqn{Q = n(n+2)\cdot\sum_{j=1}^h \hat{\rho}(j)^2/(n-j)} with \emph{n} the number of observations
and \eqn{\hat{\rho}(j)} the autocorrelation coefficient in the sample when the lag is \emph{j}.
\code{Box.Ljung.Test} computes \eqn{Q} and returns the p-values graph with lag \emph{j}.
}
\examples{
# Example for CRAN checks:
# Executable in < 5 sec

z <- rnorm(500)
Box.Ljung.Test(z, lag = 15)
ts.diag(z)
}
\references{
For more information on theoretical foundations and estimation methods see

\insertRef{brockwell2002introduction}{lsts}

\insertRef{ljung1978measure}{lsts}
}
\seealso{
\code{\link{periodogram}}, \code{\link[graphics]{persp}}
}
