% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_kalman.R
\name{LS.kalman}
\alias{LS.kalman}
\title{Kalman filter for locally stationary processes}
\usage{
LS.kalman(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{ls_kalman()} replaces this function
}
