% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_summary.R
\name{LS.summary}
\alias{LS.summary}
\title{Summary for Locally Stationary Time Series}
\usage{
LS.summary(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{ls_summary()} replaces this function
}
