% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_whittle_loglik.R
\name{ls_whittle_loglik}
\alias{ls_whittle_loglik}
\title{Locally Stationary Whittle log-likelihood Function}
\usage{
ls_whittle_loglik(
  x,
  series,
  order = c(p = 0, q = 0),
  ar_order = NULL,
  ma_order = NULL,
  sd_order = NULL,
  d_order = NULL,
  include_d = FALSE,
  w = NULL,
  s = NULL,
  include_taper = TRUE
)
}
\arguments{
\item{x}{(type: numeric) parameter vector.}

\item{series}{(type: numeric) univariate time series.}

\item{order}{(type: numeric) vector corresponding to \code{ARMA} model
entered.}

\item{ar_order}{(type: numeric) AR polimonial order.}

\item{ma_order}{(type: numeric) MA polimonial order.}

\item{sd_order}{(type: numeric) polinomial order noise scale factor.}

\item{d_order}{(type: numeric) \code{d} polinomial order, where \code{d} is
the \code{ARFIMA} parameter.}

\item{include_d}{(type: numeric) logical argument for \code{ARFIMA} models.
If \code{include_d=FALSE} then the model is an ARMA process.}

\item{w}{(type: numeric) value corresponding to the length of the window to
compute periodogram. If \code{w=NULL} then the function will use
\eqn{w = \text{trunc}(n^{0.8})}, see Dahlhaus (1998) where \eqn{n} is the
length of the \code{y} vector.}

\item{s}{(type: numeric) value corresponding to the lag with which will go
taking the blocks or windows.}

\item{include_taper}{(type: logical) logical argument that by default is
\code{TRUE}. See \code{\link{periodogram}}.}
}
\description{
This function computes Whittle estimator for LS-ARMA and
LS-ARFIMA models, in data with mean zero. If mean is not zero, then it is
subtracted to data.
}
\details{
The estimation of the time-varying parameters can be carried out by means of
the Whittle log-likelihood function proposed by Dahlhaus (1997),
\deqn{L_n(\theta) = \frac{1}{4\pi}\frac{1}{M} \int_{-\pi}^{\pi}
\bigg\{log f_{\theta}(u_j,\lambda) +
\frac{I_N(u_j, \lambda)}{f_{\theta}(u_j,\lambda)}\bigg\}\,d\lambda}
where \eqn{M} is the number of blocks, \eqn{w} the length of the series per
block, \eqn{n =s(M-1)+w}, \eqn{s} is the shift from block to block,
\eqn{u_j =t_j/n}, \eqn{t_j =s(j-1)+w/2}, \eqn{j =1,\ldots,M} and
\eqn{\lambda} the Fourier frequencies in the block
(\eqn{2\,\pi\,k/w}, \eqn{k = 1,\ldots, w}).
}
\references{
For more information on theoretical foundations and estimation methods see

\insertRef{brockwell2002introduction}{lsts2}

\insertRef{palma2010efficient}{lsts2}
}
\seealso{
\code{\link[stats]{nlminb}}, \code{\link{ls_kalman}}
}
