% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_whittle_loglik_sd.R
\name{ls_whittle_loglik_sd}
\alias{ls_whittle_loglik_sd}
\title{Locally Stationary Whittle Log-likelihood sigma}
\usage{
ls_whittle_loglik_sd(
  x,
  series,
  order = c(p = 0, q = 0),
  ar_order = NULL,
  ma_order = NULL,
  sd_order = NULL,
  d_order = NULL,
  include_d = FALSE,
  N = NULL,
  S = NULL,
  include_taper = TRUE,
  theta_par = numeric()
)
}
\arguments{
\item{x}{(type: numeric) parameter vector.}

\item{series}{(type: numeric) univariate time series.}

\item{order}{(type: numeric) vector corresponding to \code{ARMA} model
entered.}

\item{ar_order}{(type: numeric) AR polimonial order.}

\item{ma_order}{(type: numeric) MA polimonial order.}

\item{sd_order}{(type: numeric) polinomial order noise scale factor.}

\item{d_order}{(type: numeric) \code{d} polinomial order, where \code{d} is
the \code{ARFIMA} parameter.}

\item{include_d}{(type: numeric) logical argument for \code{ARFIMA} models.
If \code{include_d=FALSE} then the model is an ARMA process.}

\item{N}{(type: numeric) value corresponding to the length of the window to
compute periodogram. If \code{N=NULL} then the function will use
\eqn{N = \text{trunc}(n^{0.8})}, see Dahlhaus (1998) where \eqn{n} is the
length of the \code{y} vector.}

\item{S}{(type: numeric) value corresponding to the lag with which will go
taking the blocks or windows.}

\item{include_taper}{(type: logical) logical argument that by default is
\code{TRUE}. See \code{\link{periodogram}}.}

\item{theta_par}{(type: numeric) vector with the known parameters of the
model.}
}
\description{
This function calculates log-likelihood with known \eqn{\theta},
through \code{ls_whittle_loglik} function.
}
\details{
This function computes \code{\link{ls_whittle_loglik}} with \code{x} as
\code{x = c(theta_par, x)}.
}
