% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts-package.R
\docType{package}
\name{lsts-package}
\alias{lsts}
\alias{lsts-package}
\title{lsts: Locally Stationary Time Series}
\description{
A set of functions that allows stationary analysis and locally stationary time series analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://pachamaltese.github.io/lsts}
  \item Report bugs at \url{https://github.com/pachamaltese/lsts/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Olea \email{raolea@uc.cl} [copyright holder]

Authors:
\itemize{
  \item Wilfredo Palma \email{wilfredo@mat.puc.cl} [copyright holder]
  \item Pilar Rubio \email{parubio@puc.cl}
  \item Mauricio Vargas \email{mvargas@dcc.uchile.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
}

}
\keyword{internal}
