% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts_diagnostics.R
\name{lsts_diagnostics}
\alias{lsts_diagnostics}
\title{Diagnostic Plots for Time Series fits}
\usage{
lsts_diagnostics(x, lag = 10, cex = 0.5)
}
\arguments{
\item{x}{(type: numeric) residuals of the fitted time series model.}

\item{lag}{(type: numeric) maximum lag at which to calculate the acf and
Ljung-Box test. By default set to 10.}

\item{cex}{(type: numeric) optional argument of \code{\link[graphics]{par}}.
By default set to 0.5.}
}
\value{
** COMPLETE **
}
\description{
Plot time-series diagnostics.
}
\details{
This function plot the residuals, the autocorrelation function of the
residuals (ACF) and the p-values of the Ljung-Box Test for all lags up to
\code{lag}.
}
\examples{
# Examples for CRAN checks:
# Executable in < 5 sec

z <- rnorm(500)
lsts_lbtp(z, lag = 15)
lsts_diagnostics(z)

}
\seealso{
\code{\link{lsts_lbtp}}
}
