% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts_lbtp.R
\name{lsts_lbtp}
\alias{lsts_lbtp}
\title{Ljung-Box Test Plot}
\usage{
lsts_lbtp(z, lag = 10, main = NULL)
}
\arguments{
\item{z}{COMPLETE}

\item{lag}{COMPLETE}

\item{main}{COMPLETE}
}
\value{
** COMPLETE **
}
\description{
Plots the p-values Ljung-Box test.
}
\details{
The Ljung-Box test is used to check if exists autocorrelation in a time
series. The statistic is
\deqn{Q = n(n+2)\cdot\sum_{j=1}^h \hat{\rho}(j)^2/(n-j)} with \emph{n} the
number of observations and \eqn{\hat{\rho}(j)} the autocorrelation
coefficient in the sample when the lag is \emph{j}. \code{lsts_lbtp}
computes \eqn{Q} and returns the p-values graph with lag \emph{j}.
}
\examples{
z <- rnorm(500)
lsts_lbtp(z, lag = 15)
lsts_diagnostics(z)
}
\references{
For more information on theoretical foundations and estimation methods see

\insertRef{brockwell2002introduction}{lsts}

\insertRef{ljung1978measure}{lsts}
}
\seealso{
\code{\link{lsts_periodogram}}, \code{\link[graphics]{persp}}
}
