% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts_summary.R
\name{lsts_summary}
\alias{lsts_summary}
\title{Summary for Locally Stationary Time Series}
\usage{
lsts_summary(object)
}
\arguments{
\item{object}{(type: COMPLETE) \code{\link{lsts_wl}} function}
}
\value{
A list with the following components:
\item{summary }{a resume table with estimate, std. error, z-value and p-value
of the model.}
\item{aic }{AIC of the model.}
\item{npar }{number of parameters in the model.}
}
\description{
Function used to produce summaries of the results to Whittle
estimator to Locally Stationary Time Series (\code{\link{lsts_wl}}
function).
}
\details{
** COMPLETE **
}
\examples{
# Examples for CRAN checks:
# Executable in < 5 sec
}
