% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts-package.R
\docType{data}
\name{malleco}
\alias{malleco}
\title{Average Araucaria Araucana Tree Ring Width}
\format{
A time series object with 734 elements
}
\description{
A ts object containing average annual ring width measured in
milimiters for different Araucaria Araucana trees in the Malleco Region
(Chile). The years of observation in this data cover the period 1242-1975.
}
\author{
National Oceanic and Atmospheric Administration (NOAA)
}
\keyword{data}
