% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_periodogram_blocks.R
\name{smooth_periodogram_blocks}
\alias{smooth_periodogram_blocks}
\title{Smooth Periodogram by Blocks}
\usage{
smooth_periodogram_blocks(y, x = NULL, N = NULL, S = NULL,
  p = 0.25, spar.freq = 0, spar.time = 0, theta = 0, phi = 0,
  xlim = NULL, ylim = NULL, zlim = NULL, ylab = "Time",
  palette.col = NULL)
}
\arguments{
\item{y}{(type: numeric) data vector}

\item{x}{(type: numeric) optional vector, if \code{x = NULL} then the
function uses \eqn{(1,\ldots,n)} where \code{n} is the length of \code{y}.
More details in \code{y} argument from \code{\link[graphics]{persp}}
function.}

\item{N}{(type: numeric) value corresponding to the length of the window to
compute periodogram.
If \code{N=NULL} then the function will use
\eqn{N = \textmd{trunc}(n^{0.8})}, see
\insertCite{dahlhaus1998optimal;textual}{lsts} where \eqn{n} is the length of
the \code{y} vector.}

\item{S}{(type: numeric) value corresponding to the lag with which will be
taking the blocks or windows to calculate the periodogram.}

\item{p}{(type: numeric) value used if it is desired that \code{S} is
proportional to \code{N}. By default \code{p=0.25}, if \code{S} and \code{N}
are not entered.}

\item{spar.freq}{(type: numeric) smoothing parameter, typically (but not
necessarily) in \eqn{(0,1]}.}

\item{spar.time}{(type: numeric) smoothing parameter, typically (but not
necessarily) in \eqn{(0,1]}.}

\item{theta}{(type: numeric) angle defining the viewing direction, gives the
azimuthal direction.}

\item{phi}{(type: numeric) angle defining the viewing direction, gives the
colatitude.}

\item{xlim}{(type: numeric) x-axis limits, \code{NULL} by default and
optional parameter.}

\item{ylim}{(type: numeric) y-axis limits, \code{NULL} by default and
optional parameter.}

\item{zlim}{(type: numeric) z-axis limits, \code{NULL} by default and
optional parameter.}

\item{ylab}{(type: character) title for the y-axis. By default is
\code{ylab='Time'} and is an optional
parameter.}

\item{palette.col}{(type: character) colors palette.}
}
\value{
** COMPLETE **
}
\description{
Plots in 3D the smoothing periodogram of a time series, by
blocks or windows.
}
\details{
The number of windows of the function is \eqn{M = \textmd{trunc}((n-N)/S+1)},
where \code{\link[base]{trunc}} truncates de entered value and \emph{n} is
the length of the vector \code{y}. All windows are of the same length
\code{N}, if this value isn't entered by user then is computed as
\eqn{N=\textmd{trunc}(n^{0.8})} (Dahlhaus).

\code{smooth_periodogram_blocks} computes the periodogram in each of the
\emph{M} windows and then smoothes it two times with
\code{\link[stats]{smooth.spline}} function; the first time using
\code{spar.freq} parameter and the second time with \code{spar.time}. These
windows overlap between them.

The surface is then viewed by looking at the origin from a direction defined
by \code{theta} and \code{phi}. If \code{theta} and \code{phi} are both zero
the viewing direction is directly down the negative y axis. Changing
\code{theta} will vary the azimuth and changing \code{phi} the colatitude.
}
\examples{
# Examples for CRAN checks:
# Executable in < 5 sec

set.seed(2015)

ts.sim <- arima.sim(
 n = 1000, model = list(order = c(2, 0, 0), ar = c(1.3, -0.6))
)

smooth_periodogram_blocks(
  y = ts.sim, spar.freq = .9, spar.time = .9, theta = 30, phi = 0,
  N = 500, S = 100, ylab = 'Time'
)

smooth_periodogram_blocks(
  y = ts.sim, spar.freq = .9, spar.time = .9, theta = 0, phi = 0,
  N = 500, S = 100, ylab = 'Time'
)

smooth_periodogram_blocks(
  y = ts.sim, spar.freq = .9, spar.time = .9, theta = 90, phi = 0,
  N = 500, S = 100, ylab = 'Time'
)

smooth_periodogram_blocks(
  y = ts.sim, spar.freq = .9, spar.time = .9, theta = 45, phi = 15,
  N = 500, S = 100, ylab = 'Time'
)

smooth_periodogram_blocks(
  y = ts.sim, spar.freq = .9, spar.time = .9, theta = 45, phi = 15,
  N = 500, S = 100, ylab = 'Time',
  palette.col = gray(level = seq(0.2, 0.9, 0.1))
)
}
\references{
For more information on theoretical foundations and estimation methods see

\insertRef{dahlhaus1997fitting}{lsts}

\insertRef{dahlhaus1998optimal}{lsts}
}
\seealso{
\code{\link{arima.sim}}
}
