% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.diag.R
\name{ts.diag}
\alias{ts.diag}
\title{Diagnostic Plots for Time Series fits}
\usage{
ts.diag(x, lag = 10, cex = 0.5)
}
\arguments{
\item{x}{(type: numeric) residuals of the fitted time series model.}

\item{lag}{(type: numeric) maximum lag at which to calculate the acf and
Ljung-Box test. By default set to 10.}

\item{cex}{(type: numeric) optional argument of \code{\link[graphics]{par}}.
By default set to 0.5.}
}
\value{
** COMPLETE **
}
\description{
Plot time-series diagnostics.
}
\details{
This function plot the residuals, the autocorrelation function of the
residuals (ACF) and the p-values of the Ljung-Box Test for all lags up to
\code{lag}.
}
\examples{
z <- rnorm(500)
Box.Ljung.Test(z, lag = 15)
ts.diag(z)
}
\seealso{
\code{\link{Box.Ljung.Test}}
}
