% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_smooth_periodogram.R
\name{block.smooth.periodogram}
\alias{block.smooth.periodogram}
\title{Smooth Periodogram by Blocks}
\usage{
block.smooth.periodogram(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{block_smooth_periodogram()} replaces this function
}
