% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_ljung_test.R
\name{Box.Ljung.Test}
\alias{Box.Ljung.Test}
\title{Ljung-Box Test Plot}
\usage{
Box.Ljung.Test(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{box_ljung_test()} replaces this function
}
