% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_whittle_loglik.R
\name{LS.whittle.loglik}
\alias{LS.whittle.loglik}
\title{Locally Stationary Whittle log-likelihood Function}
\usage{
LS.whittle.loglik(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{ls_whittle_loglik()} replaces this function
}
