% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_whittle_loglik_sd.R
\name{LS.whittle.loglik.sd}
\alias{LS.whittle.loglik.sd}
\title{Whittle estimator to Locally Stationary Time Series}
\usage{
LS.whittle.loglik.sd(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{ls_whittle_loglik_theta()} replaces this function
}
