% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_whittle.R
\name{LS.whittle}
\alias{LS.whittle}
\title{Whittle estimator to Locally Stationary Time Series}
\usage{
LS.whittle(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{ls_summary()} replaces this function
}
