% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts_hm.R
\name{lsts_hm}
\alias{lsts_hm}
\title{Hessian Matrix}
\usage{
lsts_hm(f, x0, ...)
}
\arguments{
\item{f}{(type: numeric) name of function that defines log likelihood
(or negative of it).}

\item{x0}{(type: numeric) scalar or vector of parameters that give the point
at which you want the hessian estimated (usually will be the mle).}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
An \code{n x n} matrix of 2nd derivatives, where \emph{n} is the length of
\code{x0}.
}
\description{
Plots in 3D the smoothing periodogram of a time series,
by blocks or windows.
}
\details{
Computes numerical approximation to Hessian of \code{f}, evaluated at
\code{x0}.
Usually needs to pass additional parameters (e.g. data).  N.B. this uses no
numerical sophistication.
}
\examples{
# Variance of the maximum likelihood estimator for mu parameter in
# gaussian data

loglik <- function(series, x, sd = 1) {
  -sum(log(dnorm(series, mean = x, sd = sd)))
}

n <- 500

set.seed(1776)
series <- rnorm(500, mean = 10, sd = 2)

sqrt(c(var(series) / n, diag(solve(lsts_hm(
  f = loglik, x = mean(series), series = series,
  sd = sd(series)
)))))

}
\seealso{
\code{\link{arima.sim}}
}
