% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts-package.R
\docType{data}
\name{malleco}
\alias{malleco}
\title{Malleco (PROVIDE DETAILS)}
\format{
A time series object with 734 elements
}
\source{
Datos Ricardo Olea
}
\usage{
malleco
}
\description{
A ts object... EXPLAIN
}
\keyword{datasets}
