% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsts-package.R
\docType{data}
\name{malleco}
\alias{malleco}
\title{Average Araucaria Araucana Tree Ring Width}
\format{
A time series object with 734 elements
}
\source{
Datos Ricardo Olea

National Oceanic and Atmospheric Administration (NOAA)
}
\usage{
malleco
}
\description{
A ts object containing average annual ring width measured in milimiters 
for different Araucaria Araucana trees in the Malleco Region (Chile). The years 
of observation in this data cover the period 1242-1975.
}
\details{
A ts object... EXPLAIN
}
\keyword{datasets}
