% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.periodogram.R
\name{smooth.periodogram}
\alias{smooth.periodogram}
\title{Smoothing periodogram}
\usage{
smooth.periodogram(y, plot = TRUE, spar = 0)
}
\arguments{
\item{y}{(type: numeric) data vector.}

\item{plot}{(type: logical) logical argument which allows to plot the
periodogram.}

\item{spar}{(type: numeric) smoothing parameter, typically (but not
necessarily) in \eqn{(0,1].}}
}
\value{
** COMPLETE **
}
\description{
This function returns the smoothing periodogram of a stationary
time serie, its plot and its Fourier frequency.
}
\details{
\code{smooth.periodogram} computes the periodogram from \code{y} vector and
then smooth it with \emph{smoothing spline} method, which basically
approximates a curve using a cubic spline (see more details in
\code{\link[stats]{smooth.spline}}). \eqn{\lambda} is the Fourier frequency
obtained through \code{\link{periodogram}}.

It must have caution with the minimum length of \code{y}, because
\code{smooth.spline} requires the entered vector has at least length 4 and
the length of \code{y} does not equal to the length of the data of the
periodogram that \code{smooth.spline} receives.

If it presents problems with tol (\bold{tol}erance), see
\code{\link[stats]{smooth.spline}}.
}
\examples{
# AR(1) simulated

set.seed(1776)

ts.sim <- arima.sim(n = 1000, model = list(order = c(1, 0, 0), ar = 0.7))

periodogram(ts.sim)

aux <- smooth.periodogram(ts.sim, plot = FALSE, spar = .7)

lines(aux$smooth.periodogram ~ aux$lambda, lwd = 2, col = "orange")

lines(fdensity(ar = 0.7, lambda = aux$lambda)~aux$lambda, col = "red")

}
\seealso{
\code{\link{smooth.spline}}, \code{\link{periodogram}}
}
