% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_periodogram.R
\name{smooth.periodogram}
\alias{smooth.periodogram}
\title{Smoothing periodogram}
\usage{
smooth.periodogram(...)
}
\arguments{
\item{...}{old parameters}
}
\description{
\code{smooth_periodogram()} replaces this function
}
